removed method applyPluginIfAbsent
This method was simply not necessary: The javadoc of PluginContainer.apply
says: "Applies a plugin to the project. This usually means that the plugin
uses the project API to add and modify the state of the project.
This method can be called an arbitrary number of times for a particular
plugin type. The plugin will be actually used only the first time this
method is called."